Remove templated pjsip.conf.j2
The default pjsip.conf.j2 was causing problems with people that wanted
to use this image as a base for their own image and were copying their
own pjsip.conf into the /etc/asterisk directory without needing it to
be templated. This caused the template file to be processed on entry,
overwriting their config, which is obnoxious. The default pjsip.conf.j2
provided in this image was not really applicable in most scenarios,
so its best that we just remove it.